Feat: Add release command for `cargo-release` support
This commit adds a new release command to make version management easier with `cargo-release`.
The release command allows users to release a new version of the library with a single command,
without having to manually update version numbers in various places.
This is achieved through integration with the `cargo-release` tool.